refactor: batch 및 core 모듈 리팩토링 및 미사용 코드 제거
