+++ stockwellness-batch/src/main/java/org/stockwellness/StockwellnessBatchApplication.java
@@ -4,8 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
-@SpringBootApplication
-@ConfigurationPropertiesScan
+@SpringBootApplication(scanBasePackages = "org.stockwellness")
+@ConfigurationPropertiesScan(basePackages = "org.stockwellness")
 public class StockwellnessBatchApplication {
 
     public static void main(String[] args) {
