+++ stockwellness-batch/src/main/java/org/stockwellness/batch/domain/service/TechnicalIndicatorCalculator.java
@@ -1,7 +1,8 @@
-package org.stockwellness.domain.stock.price;
+package org.stockwellness.batch.domain.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.stockwellness.domain.stock.price.TechnicalIndicators;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
