+++ stockwellness-batch/src/test/java/org/stockwellness/batch/domain/service/TechnicalIndicatorCalculatorTest.java
@@ -1,10 +1,11 @@
-package org.stockwellness.domain.stock.price;
+package org.stockwellness.batch.domain.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import org.stockwellness.domain.stock.price.TechnicalIndicators;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
