+++ stockwellness-core/src/testFixtures/java/org/stockwellness/fixture/StockFixture.java
@@ -3,7 +3,6 @@ package org.stockwellness.fixture;
 
 import org.stockwellness.domain.stock.Currency;
 import org.stockwellness.domain.stock.MarketType;
-import org.stockwellness.domain.stock.SectorCode;
 import org.stockwellness.domain.stock.Stock;
 import org.stockwellness.domain.stock.StockStatus;
 
@@ -70,35 +69,6 @@ public class StockFixture {
                 "999",
                 "미분류",
                 StockStatus.DELISTED
-        );
-    }
-
-    /**
-     * 커스텀 종목 생성용 범용 메서드
-     */
-    public static Stock create(String ticker, String name, MarketType marketType) {
-        return Stock.of(
-                ticker,
-                "STD_" + ticker,
-                name,
-                marketType,
-                Currency.KRW,
-                SectorCode.UNKNOWN.getCode(),
-                SectorCode.UNKNOWN.getLabel(),
-                StockStatus.ACTIVE
-        );
-    }
-
-    public static Stock createWithSector(String ticker, String name, MarketType marketType, SectorCode sector) {
-        return Stock.of(
-                ticker,
-                "STD_" + ticker,
-                name,
-                marketType,
-                Currency.KRW,
-                sector.getCode(),
-                sector.getLabel(),
-                StockStatus.ACTIVE
         );
     }
 }
